refactor(glm5): split bare call tokenizer
Ultraworked with [omo](https://github.com/code-yeongyu/oh-my-openagent)
