+++ src/core/protocols/glm5-bare-tool-call.ts
@@ -7,18 +7,10 @@ import {
 } from "../utils/prototype-sensitive-keys";
 import { coerceToolCallInput } from "../utils/tool-call-coercion";
 import { getToolInputPropertyNames } from "../utils/tool-call-object-schema";
+import { scanBareNamedArguments } from "./glm5-bare-tool-call-tokenizer";
 
 const MAX_BARE_TOOL_CALL_LENGTH = 102_400;
-const MAX_BARE_TOOL_CALL_NESTING_DEPTH = 256;
-const MAX_BARE_TOOL_CALL_ARGUMENTS = 1024;
 const JSON_NUMBER_RE = /^-?\d+(?:\.\d+)?(?:[eE][+-]?\d+)?$/;
-
-type JsonishQuote = '"' | "'";
-
-interface ScannedArgument {
-  key: string;
-  rawValue: string;
-}
 
 interface ParsedValue {
   ok: true;
@@ -26,122 +18,6 @@ interface ParsedValue {
 }
 
 const INVALID_VALUE = { ok: false } as const;
-
-function appendScannedArgument(options: {
-  arguments_: ScannedArgument[];
-  body: string;
-  end: number;
-  equals: number;
-  start: number;
-}): boolean {
-  if (
-    options.equals < options.start ||
-    options.arguments_.length >= MAX_BARE_TOOL_CALL_ARGUMENTS
-  ) {
-    return false;
-  }
-
-  const key = options.body.slice(options.start, options.equals).trim();
-  const rawValue = options.body.slice(options.equals + 1, options.end).trim();
-  if (!(key && rawValue)) {
-    return false;
-  }
-  options.arguments_.push({ key, rawValue });
-  return true;
-}
-
-/**
- * Split only top-level `key=value` pairs. Quotes and JSON-ish containers own
- * their commas and equals signs; malformed or incomplete structure is never
- * completed by this fallback.
- */
-// biome-ignore lint/complexity/noExcessiveCognitiveComplexity: Keeping scanner state transitions together makes its fail-closed grammar auditable.
-function scanNamedArguments(body: string): ScannedArgument[] | null {
-  if (body.trim().length === 0) {
-    return [];
-  }
-
-  const arguments_: ScannedArgument[] = [];
-  const stack: ("[" | "{")[] = [];
-  let quote: JsonishQuote | null = null;
-  let escaping = false;
-  let segmentStart = 0;
-  let equals = -1;
-
-  for (let index = 0; index < body.length; index += 1) {
-    const char = body.charAt(index);
-    if (quote !== null) {
-      if (escaping) {
-        escaping = false;
-      } else if (char === "\\") {
-        escaping = true;
-      } else if (char === quote) {
-        quote = null;
-      }
-      continue;
-    }
-
-    if (char === '"' || char === "'") {
-      quote = char;
-      continue;
-    }
-    if (char === "{" || char === "[") {
-      stack.push(char);
-      if (stack.length > MAX_BARE_TOOL_CALL_NESTING_DEPTH) {
-        return null;
-      }
-      continue;
-    }
-    if (char === "}" || char === "]") {
-      const expected = char === "}" ? "{" : "[";
-      if (stack.pop() !== expected) {
-        return null;
-      }
-      continue;
-    }
-    if (char === "(" || char === ")") {
-      return null;
-    }
-    if (stack.length > 0) {
-      continue;
-    }
-    if (char === "=") {
-      if (equals !== -1) {
-        return null;
-      }
-      equals = index;
-      continue;
-    }
-    if (char === ",") {
-      if (
-        !appendScannedArgument({
-          arguments_,
-          body,
-          end: index,
-          equals,
-          start: segmentStart,
-        })
-      ) {
-        return null;
-      }
-      segmentStart = index + 1;
-      equals = -1;
-    }
-  }
-
-  if (quote !== null || escaping || stack.length > 0) {
-    return null;
-  }
-  return appendScannedArgument({
-    arguments_,
-    body,
-    end: body.length,
-    equals,
-    start: segmentStart,
-  })
-    ? arguments_
-    : null;
-}
 
 function parseQuotedValue(
   rawValue: string
@@ -257,7 +133,7 @@ export function parseGlm5AnchoredBareToolCall(options: {
   if (!anchored) {
     return null;
   }
-  const scanned = scanNamedArguments(anchored.body);
+  const scanned = scanBareNamedArguments(anchored.body);
   if (!scanned) {
     return null;
   }
